Split test utils to separate file
